requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.2
bidict==0.23.1
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.8
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
idna==3.11
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
python-decouple==3.8
python-engineio==4.12.3
python-socketio==5.15.0
requests==2.32.5
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.4
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
whitenoise==6.11.0
wsproto==1.3.2
